separate files per each parser
