+++ parser/errorformat_test.go
@@ -1,0 +1,16 @@
+package parser
+
+import "testing"
+
+func TestNewErrorformatParserString(t *testing.T) {
+	in := []string{`%f:%l:%c:%m`, `%-G%.%#`}
+
+	got, err := NewErrorformatParserString(in)
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	if len(got.efm.Efms) != len(in) {
+		t.Errorf("NewErrorformatParserString: len: got %v, want %v", len(got.efm.Efms), len(in))
+	}
+}
